feat: profileEdit 업데이트

+++ src/main/java/com/example/crud/data/member/dto/MemberResponseDto.java
@@ -8,7 +8,7 @@ import lombok.*;
 @NoArgsConstructor
 @Builder
 public class MemberResponseDto {
-    private int number;
+    private Long number;
     private String password;
     private String name;
     private String email;
